Added Section, Cleaned Code
I added in Jan-Feb, but I need to add in the tags for people groups. I'm still a bit confused on that, so I'm going to add those in after asking Emma about them tomorrow (1/27).
I also cleaned up the file in general, because some of the indentations were a bit off.